review feedback: split up share merging and splitting into their own files
